Move all `Google Pay`-related services under the `GooglePayPayment` namespace
