+++ config/validation/Pay.xml
@@ -56,14 +56,14 @@
             </constraint>
         </property>
         <property name="googlePayToken">
-            <constraint name="CommerceWeavers\SyliusTpayPlugin\Api\Validator\Constraint\NotBlankIfGatewayConfigTypeEquals">
-                <option name="paymentMethodType">google_pay</option>
+            <constraint name="CommerceWeavers\SyliusTpayPlugin\Api\Validator\Constraint\NotBlankIfGatewayNameEquals">
+                <option name="gatewayName">tpay_google_pay</option>
                 <option name="fieldRequiredErrorMessage">commerce_weavers_sylius_tpay.shop.pay.google_pay_token.required</option>
                 <option name="groups">
                     <value>commerce_weavers_sylius_tpay:shop:order:pay</value>
                 </option>
             </constraint>
-            <constraint name="CommerceWeavers\SyliusTpayPlugin\Validator\Constraint\EncodedGooglePayToken">
+            <constraint name="CommerceWeavers\SyliusTpayPlugin\GooglePayPayment\Validator\Constraint\EncodedGooglePayToken">
                 <option name="groups">
                     <value>commerce_weavers_sylius_tpay:shop:order:pay</value>
                 </option>
