+++ src/GooglePayPayment/Validator/Constraint/EncodedGooglePayToken.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CommerceWeavers\SyliusTpayPlugin\Validator\Constraint;
+namespace CommerceWeavers\SyliusTpayPlugin\GooglePayPayment\Validator\Constraint;
 
 use Symfony\Component\Validator\Constraint;
