+++ templates/shop/cart/complete/_google_pay.html.twig
@@ -1,6 +1,4 @@
-{% if payment is not null and
-    cw_tpay_get_gateway_config_value(payment.method.gatewayConfig, 'type') == constant('CommerceWeavers\\SyliusTpayPlugin\\Tpay\\PaymentType::GOOGLE_PAY')
-%}
+{% if payment is not null and payment.method.gatewayConfig.gatewayName == 'tpay_google_pay' %}
     {% set payment = order.lastCartPayment() %}
     {% set gatewayConfig = payment.method.gatewayConfig %}
     {% set productionMode = gatewayConfig.config['production_mode'] ?? true %}
