+++ templates/shop/order/pay/_google_pay_regulations.html.twig
@@ -1,7 +1,4 @@
-{% if cw_tpay_get_gateway_config_value(method.gatewayConfig, 'type') ==
-    constant('CommerceWeavers\\SyliusTpayPlugin\\Tpay\\PaymentType::GOOGLE_PAY')
-    and form.vars.value.state != 'cart'
-%}
+{% if method.gatewayConfig.gatewayName == 'tpay_google_pay' and form.vars.value.state != 'cart' %}
     <div class="ui stackable grid" style="display: none;" data-payment-details="google_pay">
         <div class="ui stackable grid">
             <div class="ui column eight wide">
